+++ src/agents/__init__.py
@@ -1,7 +1,6 @@
 """Entry point into the agents module set"""
 
 from .base_agent import BaseAgent
-from .base_rl_agent import BaseRLAgent
 from .random_agent import RandomAgent
 from .rotating_agent import RotatingAgent
 from .baseline_wacuum_cleaner import WacuumCleaner
